feat: add group validation in domain layer - Add check for contacts using a group before deletion - Update GroupDomainService to prevent deletion of groups with contacts

+++ src/main/java/domain/group/GroupDomainService.java
@@ -4,13 +4,17 @@ import java.io.IOException;
 import java.util.HashSet;
 import java.util.Set;
 
+import domain.contact.ContactRepository;
+
 public class GroupDomainService {
 
 
     GroupRepository repository;
+    ContactRepository contactRepository;
 
-    public GroupDomainService(GroupRepository repository) {
+    public GroupDomainService(GroupRepository repository, ContactRepository contactRepository) {
         this.repository = repository;
+        this.contactRepository = contactRepository;
     }
 
     public void add(Group group) throws IOException {
@@ -24,6 +28,15 @@ public class GroupDomainService {
         if (repository.find(name).isEmpty()) {
             throw new IOException("Group with name = " + name + " doesn't exist");
         }
+        
+        // Check if any contacts are using this group
+        boolean hasContacts = contactRepository.findAll().stream()
+                .anyMatch(contact -> name.equals(contact.getGroup()));
+        
+        if (hasContacts) {
+            throw new IOException("Cannot delete group '" + name + "' as it has contacts associated with it");
+        }
+        
         repository.delete(name);
     }
 
@@ -38,8 +51,9 @@ public class GroupDomainService {
     }
 
     public void update(Group group) throws IOException {
-        if (!get(group.getName()).equals(group)) {
-            repository.save(group);
+        if (repository.find(group.getName()).isEmpty()) {
+            throw new IOException("Group with name = " + group.getName() + " does not exist.");
         }
+        repository.save(group);
     }
 }
